feat:添加静态资源，修改组件设计
