--- src/pages/Banner.vue
@@ -1,15 +1,0 @@
-<template>
-  <img src="" alt="" class="bannerImg" />
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped>
-.bannerImg {
-  display: block;
-  width: 375px;
-  height: 160px;
-  opacity: 1;
-  background-color: pink;
-}
-</style>
