--- src/pages/Swiper.vue
@@ -1,7 +1,0 @@
-<template>
-  <h1>Page</h1>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
