+++ src/style.css
@@ -1,0 +1,4 @@
+body{
+    margin: 0;
+    overflow: hidden;
+}
