+++ src/styles/reset.css
@@ -158,7 +158,3 @@ select {
 select::-ms-expand {
   display: none;
 }
-
-body {
-  overflow: hidden;
-}
